Even more line of sight + enemy improvements
added the box of sight to the hedgehog enemies and made script to chase the player.
Then the script doubles for enemy shooting as well separated by a bool.

It looks like spaghetti code to be honest, but it works and reworking it would be a pain in the ass. Maybe do it later when I know what I want to do with it. For now I like having a universal code for enemy movement/shooting.
Also the enemy hedgehogs can automatically find out who the player is as soon as the game begins.