Small changes
Did some testing from my laptop. As long as the aspect ratio is locked at 1080 there is no issue with the HUD.

Otherwise, I lowered the elasticity for the inventory since the laptop's scroll ability is way more sensitive than any computer mouse. The safest choice is to turn off elasticity, but it's not that big of an issue.